requirements.txt
flask==3.0.3
openai==1.30.1
requests==2.32.3
gspread==6.1.2
google-auth==2.29.0
reportlab==4.2.0
pillow==10.3.0
gunicorn==22.0.0
